many fixes, complex support
