+++ src/engine.cpp
@@ -7,129 +7,6 @@
 
 namespace ptMgrad {
 
-// TODO: do we need different overloads like class Value<float>, class Value<double> etc?
-template <typename T>
-class Value {
-private:
-    typedef T U;
-    U x;
-    U y;
-public:
-    Value() {};
-
-    // TODO: cannot use const directly in func param
-    template <class _X> constexpr
-    Value(const Value<_X>& _x): x(_x) {};
-
-    template <class _X> constexpr
-    Value(const Value<_X>& _x, const Value<_X>& _y): x(_x), y(_y) {}
-
-    // for scalar values
-    template <class _X> constexpr
-    Value(const _X& _x): x(_x) {}
-
-    template <class _X> constexpr
-    Value(const _X& _x, const _X& _y): x(_x), y(_y) {}
-
-    template <class _X> constexpr
-    Value& operator= (const Value<_X>& _x) {
-        // PS: "this" is not mandatory, i guess
-        this->x = _x;
-        return *this;
-    }
-
-    template <class _X> constexpr
-    Value& operator= (const _X& _x) {
-        this->x = _x;
-        return *this;
-    }
-
-    T dataX() const {
-        return x;
-    }
-
-    T dataY() const {
-        return y;
-    }
-
-    template <class _X> constexpr
-    Value& operator- () {
-        this->x = -x;
-        this->y = -y;
-        return *this;
-    }
-
-    constexpr operator bool() const {
-        return x || y;
-    }
-
-    template <class _X> constexpr
-    Value& operator+= (const Value <_X>& _x) {
-        this->x += _x.x;  // TODO: remove ".x"
-        return *this;
-    }
-
-    template <class _X> constexpr
-    Value& operator+= (const _X& _x) {
-        this->x += _x;
-        return *this;
-    }
-
-    template <class _X> constexpr
-    Value& operator-=(const Value<_X>& _x) {
-        this->x -= _x.x;
-        return *this;
-    }
-
-    template <class _X> constexpr
-    Value& operator-=(const _X& _x) {
-        this->x -= _x;
-        return *this;
-    }
-
-    template <class _X> constexpr
-    Value& operator*=(const Value<_X>& _x) {
-        this->x *= _x.x;
-        return *this;
-    }
-
-    template <class _X> constexpr
-    Value& operator*=(const _X& _x) {
-        this->x *= _x;
-        return *this;
-    }
-
-    template <class _X> constexpr
-    Value& operator/=(const Value<_X>& _x) {
-        this->x /= _x.x;
-        return *this;
-    }
-
-    template <class _X> constexpr
-    Value& operator/=(const _X& _x) {
-        this->x /= _x;
-        return *this;
-    }
-
-    // TODO: why const?
-    // TODO: Figure out why we don't need this
-    /*template <class _X> constexpr
-    bool operator<(const Value<_X>& _x) const {  // Note: return type bool
-        return x < _x.x;
-    }
-
-    template <class _X> constexpr
-    bool operator<(const _X& _x) {
-        return x < _x;
-    }*/
-
-    friend std::ostream& operator<<(std::ostream& os, const Value& _x) {
-        os << _x.x;
-        return os;
-    }
-};
-
-
 // We'll need these if we remove bool operator from Value class
 // it might return bool if we remove this
 template <class T>
@@ -159,6 +36,14 @@ operator+(const Value<T>& _x, const T&  _y) {
     return __k;
 }
 
+template <class T>
+inline
+Value<complex<T>>
+operator+(const ptMgrad::complex<T>& _x, const ptMgrad::complex<T>& _y) {
+    ptMgrad::complex<T> __k =  _x;
+    __k += _y;
+    return __k;
+}
 
 template <class T>
 inline
@@ -528,4 +413,4 @@ gt(const Value<T>& _x, const T& _y) {
     return _x > _y;
 }
 
-}  // namespace val
+}  // namespace ptMgrad
